move PostgresWriteSession
